requirements.txt
discord.py[voice]
requests
microsofttranslator
